requirements.txt
Flask==2.3.3
openai==1.10.0
google-cloud-texttospeech==2.16.3
python-dotenv==1.0.1
twilio==9.0.4
